Refactor integration tests for advise
